requirements.txt
discord.py
python-dotenv
deep-translator
requests
aiohttp
Pillow
SQLAlchemy
discord-webhook
Flask
keep_alive
rapidfuzz